unrefactor location-search
