--- src/hearings/containers/request-hearing/hearing-venue/hearing-venue.component.html
@@ -1,42 +1,0 @@
-<div class="govuk-width-container">
-  <div class="page">
-    <main class="govuk-main-wrapper govuk-main-wrapper--auto-spacing" id="main-content" role="main">
-      <div *ngIf="!findLocationFormGroup.valid && findLocationFormGroup.dirty" 
-           class="govuk-error-summary" 
-           aria-labelledby="error-summary-title" 
-           role="alert" 
-           tabindex="-1" 
-           data-module="govuk-error-summary">
-        <h2 class="govuk-error-summary__title" id="error-summary-title">
-          There is a problem
-        </h2>
-        <div class="govuk-error-summary__body">
-          <ul class="govuk-list govuk-error-summary__list">
-            <li>
-              <a href="javascript:void(0)" (click)="getLocationSearchFocus()">Enter a valid location</a>
-            </li>
-          </ul>
-        </div>
-      </div>
-      <div class="govuk-grid-row">
-        <div class="govuk-grid-column-two-thirds">      
-          <span class="govuk-caption-l">Request a hearing for Jane Smith vs DWP</span>
-          <h1 class="govuk-heading-l">
-            What are the hearing venue details?
-          </h1>
-          <div id="location-hint" class="govuk-hint govuk-!-margin-bottom-7">
-              If this is a fully remote hearing you must still select the court or tribunal which will be managing the case.
-          </div>
-          <form  [formGroup]="findLocationFormGroup" class="form" action="hearing-judge">
-            <div class="govuk-form-group">      
-              <exui-location-search-container [control]="findLocationFormGroup.controls.locationSelectedFormControl"></exui-location-search-container>
-              <div class="govuk-inset-text">
-                  You can check the venue has the required facilities or reasonable adjustments using<br> <a href="https://www.gov.uk/find-court-tribunal" target="_blank">Court Finder (opens in new tab)</a>
-              </div>      
-            </div>      
-          </form>      
-        </div>          
-      </div>           
-      </main>
-  </div>
-</div>
