--- src/hearings/containers/request-hearing/hearing-venue/hearing-venue.component.scss
@@ -1,9 +1,0 @@
-.govuk-width-container{
-    padding-left: 30px;
-    #content{
-        margin-right: 0px;
-        .govuk-back-link{
-            margin-bottom: 50px;
-        }
-    }
-}
